Add support for hint option
Inline with the SimpleForm implementation, hints (styled as per
bootstrap3 docs) can be added by passing in a string to options[:hint]

Also improved the tests a bit, which identified a bug where the label
is rendered for collections even then options[:label] isn’t supplied.
So I fixed the bug as well.